Fix issue with wrong SessionState being manipulated for message
It's possible to receive an out-of-order message belonging to different SessionState. In that case we are trying to decrypt message using all SessionStates starting from currentState and then previousStates. We have discovered that in case of failed decryption for current state and successful decryption for one of previous states, both states are saved to the store. In that case only changes in the state that this message belongs to should be preserved.
Fixed this by creating a temporary copy of a state and only saving it when decryption succeeds.

PER-1865